soccer_eventpred/modules/datamodule/soccer_datamodule.py
import pytorch_lightning as pl
import torch
from tango.common import Registrable


class SoccerDataModule(pl.LightningDataModule, Registrable):
    def train_dataloader(self) -> torch.utils.data.DataLoader:
        raise NotImplementedError

    def val_dataloader(self) -> torch.utils.data.DataLoader:
        raise NotImplementedError

    def test_dataloader(self) -> torch.utils.data.DataLoader:
        raise NotImplementedError
