pyproject.toml
[tool.poetry]
name = "soccer_eventpred"
version = "1.0.0"
description = ""
authors = ["agiats <taiga98-0809@g.ecc.u-tokyo.ac.jp>"]
readme = "README.md"
packages = [{include = "soccer_eventpred"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.26.0"
pandas = "^2.1.1"
jsonlines = "^4.0.0"
optuna = "^3.3.0"
torch = "^2.0.1"
ai2-tango = "^1.2.1"
mlflow = "^2.7.1"
fire = "^0.5.0"
scikit-learn = "^1.3.1"
pytorch-lightning = "^2.0.9"
torchmetrics = "^1.2.0"
tqdm = "^4.66.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
pysen = "^0.10.5"
flake8 = "^6.1.0"
black = "^23.9.1"
mypy = "^1.5.1"

[tool.pysen]
version = "0.10"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 88
py_version = "py310"

[tool.mypy]
python_version="3.10"
files = "soccer_eventpred"
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
pretty = true
allow_untyped_globals = true

[[tool.pysen.lint.mypy_targets]]
  paths = ["soccer_eventpred"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
